Require handlers to be passed in as a struct
This explicitly names the thing that's being passed in when initializing
the Alice application in another app. It also sets things up so that
either state backends or chat backends could be specified in the same
struct.

Backwards compatibility is not kept meaning this is a breaking change
in the API.